platform/NVMem:Fix stack overflow in _plat__NvInitFromStorage() on 64b architectures
The `UINT32 bytesRead` stack variable in _plat__NvInitFromStorage() is passed as
fourth argument to `TEE_ReadObjectData()` which has following signature:

TEE_Result TEE_ReadObjectData(TEE_ObjectHandle object, void *buffer,
			      size_t size, size_t *count)

This might cause stack overflow on 64b architectures where 64b-wide size_t
value would be written to address of 32b-wide variable on stack.

+++ platform/NVMem.c
@@ -133,7 +133,7 @@ _plat__NvInitFromStorage()
 	UINT32 i;
 	BOOL initialized;
 	UINT32 objID;
-	UINT32 bytesRead;
+	size_t bytesRead;
 	TEE_Result Result;
 
 	// Don't re-initialize.
